Begin work on new Blog section
The LGD site now has a blog section, based on the one from the Anchorage site.
It currently reuses parts of the EventGrid components that display events.
It currently has a bug where the EventGridItems that display blog Cards on the
Blog index page have their links incorrectly pointing to /events.
(This is because the EventGridItem that contains the Card component is a little
too specifically set up for events. It could be made more generic).

TODO:
- Fix the above bug so that EventGrids and EventGridItems items are built as a more generic
'Grid' and 'GridItem', or Grids allow multiple types of item (eg events, or
posts) directly)?
- Properly style the SinglePostItem pages.
- Update the headers and styling etc so that the things specific to The
Anchorage are changed or removed.
- Fix the hardcoded local ip addresses that are used to point to the CMS API