Clean up package linking
We got away with this working because `npm install` was linking these
packages into the root level `node_modules` and things resolved from there,
but its best practice for all dependencies of a project - even those from a
local workspace - to be listed in the package.json.

Often bigger projects will enforce this with linters. Other package managers
like PNPM also will force you to do it.

It wasn't strictly necessary here but I thought it was responsible for a build
error (the error was something else, but I'd already done this cleanup by the time
I realised)