Hide the overflow on imageContainer
This was causing the profile images to overlap the text content.

The issue is that `.imageContainer` on the profiles page is setting
`height` and `max-height`, and then the `.image` inside it (from
`TwoColumnImageDiv`) is setting both `height` and `min-height`.

So it was trying to control the height of the image from two different
places - the image div, and the container div. At some of the styling
sizes, the image is bigger than the container, so it overlaps.

I’ve applied an overflow:hidden to the container as a fix that seems to work well enough